requirements.txt
slack-bolt==1.23.0
openai==1.3.7
pinecone==7.2.0
python-dotenv==1.0.0
requests==2.31.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
python-pptx==0.6.23
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
PyMuPDF==1.23.8
tiktoken==0.5.2
httpx==0.25.2
certifi==2023.11.17
python-multipart==0.0.6
schedule==1.2.0 